+++ python/spglib/__init__.py
@@ -40,6 +40,7 @@ from ._version import __version__, __version_tuple__  # noqa: F401
 from .spglib import (  # noqa: F401
     MagneticSpaceGroupType,
     SpaceGroupType,
+    SpglibDataset,
     delaunay_reduce,
     find_primitive,
     get_BZ_grid_points_by_rotations,