+++ test/functional/python/test_spacegrouop_type_from_symmetry.py
@@ -31,42 +31,42 @@ def test_spacegroup_type_from_symmetry(
         if "distorted" in str(fname):
             dataset = get_symmetry_dataset(cell, symprec=1e-1)
             spgtype = get_spacegroup_type_from_symmetry(
-                dataset["rotations"],
-                dataset["translations"],
+                dataset.rotations,
+                dataset.translations,
                 lattice=lattice,
                 symprec=1e-1,
             )
-            if spgtype.number != dataset["number"]:
-                print("%d != %d in %s" % (spgtype.number, dataset["number"], fname))
+            if spgtype.number != dataset.number:
+                print("%d != %d in %s" % (spgtype.number, dataset.number, fname))
                 ref_cell = (
-                    dataset["std_lattice"],
-                    dataset["std_positions"],
-                    dataset["std_types"],
+                    dataset.std_lattice,
+                    dataset.std_positions,
+                    dataset.std_types,
                 )
                 dataset = get_symmetry_dataset(ref_cell, symprec=1e-5)
                 spgtype = get_spacegroup_type_from_symmetry(
-                    dataset["rotations"],
-                    dataset["translations"],
+                    dataset.rotations,
+                    dataset.translations,
                     lattice=lattice,
                     symprec=1e-5,
                 )
                 print(
                     "Using refined cell: %d, %d in %s"
-                    % (spgtype.number, dataset["number"], fname),
+                    % (spgtype.number, dataset.number, fname),
                 )
         else:
             dataset = get_symmetry_dataset(cell, symprec=1e-5)
             spgtype = get_spacegroup_type_from_symmetry(
-                dataset["rotations"],
-                dataset["translations"],
+                dataset.rotations,
+                dataset.translations,
                 lattice=lattice,
                 symprec=1e-5,
             )
 
-        assert spgtype.number == dataset["number"], "%d != %d in %s" % (
+        assert spgtype.number == dataset.number, "%d != %d in %s" % (
             spgtype.number,
-            dataset["number"],
+            dataset.number,
             fname,
         )
-        spgtype_ref = get_spacegroup_type(dataset["hall_number"])
+        spgtype_ref = get_spacegroup_type(dataset.hall_number)
         assert spgtype == spgtype_ref